requirements.txt
youtube-transcript-api>=0.6.0
openai>=1.0.0
notion-client>=2.0.0
python-dotenv>=1.0.0 
yt-dlp>=2023.0.0 